Add browser support
This significantly changes how tests are run. Because we need to run
tests in the browser, Playwright is used for testing. The tests now use
the Playwright snapshot feature. For every input fixture a markdown and
HTML snapshot are made both using the Node.js and the browser API.

To run tests in the browser, esbuild is used to serve a bundled version
of the browser API. This is then dynamically imported and called inside
a Playwright browser page.

Unlike the Node.js implementation, the browser implementation doesn’t
support any options. Launch options are irrelevant in the browser.
Mermaid configuration is global in the browser. SVG Minification is
probably unneeded when used in the browser, so SVGO isn’t included.

Closes #8