Support error fallback
If the user provides `errorFallback`, it will be called when mermaid
throws an error. It will receive the faulty code node, the mermaid error
message, and the vfile as arguments, and may return a node to replace
the code. If nothing is returned from the fallback, the code will be
removed instead. To keep the code as-is, simply return the it from the
fallback.

Also if no error fallback is provided, errors are now handled gracefully
using `file.fail()`.

Closes #10